+++ vault/vault.d/local/share/cook/bin/cluster-setup-nomad-proxy.sh
@@ -1,7 +1,10 @@
 #!/bin/sh
 
-# The "vault-servers" Vault group needs the "issue-nomad-client-cert" policy
+# The "consul-vault-clients" Vault group needs the "issue-nomad-client-cert" policy
 # for this to work.
+# Example commands:
+# OLD_POLICIES=$(~/vault.sh read -format=json identity/group/name/consul-vault-clients | jq -c '.data.policies' | tr -d '"[]')
+# ~/vault.sh write identity/group/name/consul-vault-clients policies=${OLD_POLICIES},issue-nomad-client-cert
 
 # shellcheck disable=SC1091
 . /root/.env.cook
@@ -34,11 +37,14 @@ sep=$'\001'
 
 # Copy over consul-template template for Nomad certs
 < "$TEMPLATEPATH/cluster-nomad.tpl.in" \
-  sed "s${sep}%%nodename%%${sep}$NODENAME${sep}g" | \
+  sed "s${sep}%%nodename%%${sep}$NODENAME${sep}g" \
   > "/mnt/templates/nomad.tpl"
 
 # Uncomment Nomad cert template in consul-template config
 sed -i '' 's/^##nomadproxy##//g' /usr/local/etc/consul-template-consul.d/consul-template-consul.hcl
+
+# Restart consul-template-consul for it to provision Nomad client certs
+service consul-template-consul restart
 
 # Copy over Nginx config for Nomad proxy
 cp "$TEMPLATEPATH/cluster-nomadproxy.conf.in" \